requirements.txt
regex
atlassian-python-api
